ゲームごとにStreamがぶんりするようになった
