+++ apps/backend/src/usecase/game/start.ts
@@ -1,6 +1,8 @@
 import { Err, Ok, Result } from "@cffnpwr/result-ts";
 import { inject, injectable } from "tsyringe";
 
+import { ChangePhaseEventPayload } from "./event";
+
 import type { IPlayerRepository } from "@/domain/repository/interface/player";
 import type { IRoomRepository } from "@/domain/repository/interface/room";
 
@@ -13,8 +15,8 @@ import {
 } from "@/error/usecase/common";
 import { PlayerNotFoundError } from "@/error/usecase/player";
 import { RoomNotFoundError } from "@/error/usecase/room";
-import { changePhaseEE } from "@/event/changePhase";
-import { ee } from "@/event/common";
+import { ee } from "@/event";
+import { EventPort } from "@/misc/event";
 import { voidType } from "@/misc/type";
 
 @injectable()
@@ -49,6 +51,11 @@ export class StartGameUseCase {
     if (roomRepoResult.isErr()) {
       return new Err(new RepositoryOperationError(roomRepoResult.unwrapErr()));
     }
+
+    const changePhaseEE: EventPort<(payload: ChangePhaseEventPayload) => void> = new EventPort(
+      `changePhase-${room.id}`,
+      ee,
+    );
 
     ee.emit(changePhaseEE, {
       eventType: "changePhase",
